requirements.txt
streamlit
pandas
numpy
plotly
biopython
fpdf
